+++ src/BongoCat.cpp
@@ -1,12 +1,14 @@
 #include "BongoCat.hpp"
 #include "CustomizeMenu.hpp"
 #include "SecretUnlocks.hpp"
+#include "EditCounterMenu.hpp"
 
 int BongoCat::m_catID = 1;
 int BongoCat::m_hatID = 1;
 int BongoCat::m_decoID = 1;
 bool BongoCat::m_hideInLevel = false;
 bool BongoCat::m_hideCounter = false;
+int BongoCat::m_pawsToUse = 3;
 
 BongoCat* BongoCat::create() {
 	auto ret = new BongoCat();
@@ -32,10 +34,14 @@ void BongoCat::addCat() {
 	m_decoID = Mod::get()->getSavedValue<int>("deco", 1);
 	m_hideInLevel = Mod::get()->getSavedValue<bool>("hideLevel", false);
 	m_hideCounter = Mod::get()->getSavedValue<bool>("hideCounter", false);
+	m_pawsToUse = Mod::get()->getSavedValue<int>("pawsMode", 3);
+
 	auto scaleX = Mod::get()->getSavedValue<float>("scaleX", 1);
 	auto newScale = Mod::get()->getSavedValue<float>("scale", 1);
 	auto posX = Mod::get()->getSavedValue<float>("posX", winSize.width - 45);
 	auto posY = Mod::get()->getSavedValue<float>("posY", 25);
+	auto bgID = Mod::get()->getSavedValue<int>("bg", 1);
+	auto fontID = Mod::get()->getSavedValue<int>("font", 0);
 
 	if (posX > winSize.width) posX = winSize.width - 45;
 	if (posY > winSize.height) posY = 25;
@@ -78,12 +84,16 @@ void BongoCat::addCat() {
 	sprDeco->setZOrder(51);
 	sprDeco->setID("deco");
 	sprDeco->setAnchorPoint(ccp(0.5, 0.01));
-	if (m_hatID == 1) sprDeco->setVisible(false);
+	if (m_decoID == 1) sprDeco->setVisible(false);
 	sprDeco->setScaleX(scaleX);
 	cat->addChild(sprDeco);
 
+	std::string bgFile;
+	if (bgID != 6) bgFile = "GJ_square0" + std::to_string(bgID) + ".png";
+	else bgFile = "square02_small.png";
 
-	auto box = CCScale9Sprite::create("GJ_square01.png");
+	auto box = CCScale9Sprite::create(bgFile.c_str());
+	if (bgID == 6) box->setOpacity(100);
 	box->setContentSize({ 160, 44 });
 	box->setScale(0.5);
 	box->setPositionY(-10);
@@ -109,11 +119,14 @@ void BongoCat::addCat() {
 	cat->addChild(menu);
 	if (m_hideCounter) menu->setVisible(false);
 
+	auto fontFile = EditCounterMenu::getFontByID(fontID);
+	auto scaleFactor = fontID == 0 ? 0.6 : 0.45;
+
 	auto amount = Mod::get()->getSavedValue<int>("count", 0);
-	auto count = CCLabelBMFont::create(std::to_string(amount).c_str(), "goldFont.fnt");
-	count->setAnchorPoint(ccp(0, 0.5));
+	auto count = CCLabelBMFont::create(std::to_string(amount).c_str(), fontFile.c_str());
+	count->setAnchorPoint(fontID <= 2 ? ccp(0, 0.5) : ccp(0, 0.6));
 	count->setPosition(ccp(-36, -9));
-	count->limitLabelWidth(52.f, 0.6f, 0.2f);
+	count->limitLabelWidth(52.f, scaleFactor, 0.2f);
 	cat->addChild(count);
 	if (m_hideCounter) count->setVisible(false);
 }
@@ -143,8 +156,19 @@ bool BongoCat::ccTouchBegan(CCTouch* touch, CCEvent* event) {
 	SecretUnlocks::m_clicksThisSession++;
 	SecretUnlocks::checkSessionUnlocks();
 
-	setFrame(m_lastPaw ? 2 : 3);
-	m_lastPaw = !m_lastPaw;
+	if (m_pawsToUse == 3) {
+		setFrame(m_lastPaw ? 2 : 3);
+		m_lastPaw = !m_lastPaw;
+	}
+
+	if (m_pawsToUse == 2) {
+		setFrame(2);
+	}
+
+	if (m_pawsToUse == 1) {
+		setFrame(3);
+	}
+
 	return true;
 }
 
@@ -210,9 +234,20 @@ void BongoCat::setToTop(float dt) {
 		}
 	}
 
-	if (runningScene->getHighestChildZ() == this->getZOrder()) return;
+	// Changing Bongo Cat z-order down again to only be one above the current highest node at all times.
+	// Some menus with set z-orders would appear behind their popup if Bongo Cat previously set it's ID to be above them. 
+	// With popups appearing at the highest z-order, they then ended up above their submenus.
+	// Also feel like this is just better, otherwise any popup would boost the z order more and more lol
+	auto children = runningScene->getChildren();
+	auto highest = static_cast<CCNode*>(children->objectAtIndex(children->count() - 1));
 
-	this->setZOrder(runningScene->getHighestChildZ() + 1);
+	if (highest->getID() == "natrium.bongo_cat/BongoCat") {
+		auto secondHighest = static_cast<CCNode*>(children->objectAtIndex(children->count() - 2));
+		this->setZOrder(secondHighest->getZOrder() + 1);
+	}
+	else {
+		this->setZOrder(highest->getZOrder() + 1);
+	}
 }
 
 void BongoCat::onSettings(CCObject* sender) {